requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
sqlalchemy==2.0.15

mysql-connector-python==8.4.0

python-dotenv==1.0.1

Werkzeug==3.0.3